+++ tests/testthat/test-primitives.R
@@ -9,6 +9,12 @@ test_that("primitives 1D", {
    ## test round-trip back to sf
    st_as_sf() %>% expect_s3_class("sf")
 })
+
+library(maptools)
+data(wrld_simpl)
+nodes <- arc_node(PRIMITIVE(st_as_sf(wrld_simpl)))
+
+
 # 
 # test_that("primitives 2D", {
 #   error("no tests!")