+++ NEWS.md
@@ -1,6 +1,6 @@
 # dev
 
-* WIP PRIMITIVES
+* PRIMITIVES model added, including bare-nodes arc_node function to return the 3-way nodes
 
 
 # sc 0.0.1